Add a base class for the fieldset legend tag helpers
The date input, checkboxes and radios legend tag helpers each carried
their own copy of the same ProcessAsync, so they now derive from
FormGroupFieldsetLegendTagHelperBase alongside the label, hint and error
message bases. Each subclass keeps only its target elements, its tag name
and its documentation.

The base passes context.TagName to SetLegend rather than the tag name
constant, matching the sibling bases, so the duplicate legend error names
whichever element was actually used. It also snapshots the content with a
null check, which the date input version already did -- a self-closing
legend in checkboxes or radios used to throw a NullReferenceException.
