+++ src/GovUk.Frontend.AspNetCore/TagHelpers/RadiosFieldsetLegendTagHelper.cs
@@ -1,4 +1,3 @@
-using GovUk.Frontend.AspNetCore.ComponentGeneration;
 using Microsoft.AspNetCore.Razor.TagHelpers;
 
 namespace GovUk.Frontend.AspNetCore.TagHelpers;
@@ -9,44 +8,14 @@ namespace GovUk.Frontend.AspNetCore.TagHelpers;
 [HtmlTargetElement(TagName, ParentTag = RadiosFieldsetTagHelper.TagName)]
 [HtmlTargetElement(TagName, ParentTag = RadiosTagHelper.TagName)]
 [TagHelperDocumentation(ContentDescription = "The content is the HTML to use within the legend. When this element is specified directly inside the root radios element a fieldset is generated automatically.")]
-public class RadiosFieldsetLegendTagHelper : TagHelper
+public class RadiosFieldsetLegendTagHelper : FormGroupFieldsetLegendTagHelperBase
 {
     internal const string TagName = "govuk-radios-fieldset-legend";
 
-    private const string IsPageHeadingAttributeName = "is-page-heading";
-
     /// <summary>
-    /// Whether the legend also acts as the heading for the page.
+    /// Creates a <see cref="RadiosFieldsetLegendTagHelper"/>.
     /// </summary>
-    /// <remarks>
-    /// The default is <c>false</c>.
-    /// </remarks>
-    [HtmlAttributeName(IsPageHeadingAttributeName)]
-    public bool? IsPageHeading { get; set; }
-
-    /// <inheritdoc/>
-    public override async Task ProcessAsync(TagHelperContext context, TagHelperOutput output)
+    public RadiosFieldsetLegendTagHelper()
     {
-        ArgumentNullException.ThrowIfNull(context);
-        ArgumentNullException.ThrowIfNull(output);
-
-        var fieldsetContext = context.GetContextItem<FormGroupFieldsetContext2>();
-
-        var content = output.TagMode == TagMode.StartTagAndEndTag ?
-            await output.GetChildContentAsync() :
-            null;
-
-        if (output.Content.IsModified)
-        {
-            content = output.Content;
-        }
-
-        fieldsetContext.SetLegend(
-            IsPageHeading,
-            new AttributeCollection(output.Attributes),
-            html: content.Snapshot(),
-            TagName);
-
-        output.SuppressOutput();
     }
 }
